Task/cms data endpoint provider modal/plrs 306 (#483)
* Setup cMsOrgPrimarySpecialtyCounts query w/ hardcoded PAC ID

* Fix query var name & add query to frontend api

* Display groupPracticePacId field in Provider Modal

* Fix random frontend graphql query problems

* Add AdditionalSections prop to quickly show extra data in AccountModals

* Build CMS specialty counts component and nest ProviderAccountModalButton with it

* Remove hardcoded query variable

* Make query network-only

* Backtrack initial approach: displaying data in modal

* Build CMS Data modal button for provider accounts page

* Change cMsOrgPrimSpecCounts return to array of sorted [key, values] and update frontend code to match

* Decrease row padding for each count

* Adjust padding again

* Remove propTypes from prev approach

* Remove extra field from delete provider organization payload

* Fix no-op in deleteTPConnectionsAndHistory util

+++ src/backend/resolvers/mutations/organization/utils/deleteTreatmentPlanConnectionsAndHistory.js
@@ -21,18 +21,21 @@ module.exports = async ({
     .collection('organizations.treatmentPlans.history')
     .aggregate(getEnrichOrgTpHistoryTrashPipeline(organizationId), { session })
     .toArray()
-    
+
   const trashDocsToInsert = enrichedTrashDocs.map(({ _id, ...doc }) => doc)
 
-  // 3. insert trash docs
-  await db
-    .collection('trash.organizations.treatmentPlans.history')
-    .insertMany(trashDocsToInsert, { session })
+  // 3. insert trash docs only if there are any
+  if (trashDocsToInsert.length) {
+    await db
+      .collection('trash.organizations.treatmentPlans.history')
+      .insertMany(trashDocsToInsert, { session })
 
-  // 4. delete all historical documents
-  await db
-    .collection('organizations.treatmentPlans.history')
-    .deleteMany({ organizationId }, { session })
+    // 4. delete all historical documents only if they exist
+    await db
+      .collection('organizations.treatmentPlans.history')
+      .deleteMany({ organizationId }, { session })
+  }
+
 
   return enrichedTrashDocs
 }
